+++ workspace/src/pages/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Pages Index
+ * 
+ * Export all page components for routing.
+ */
+
+export { UploadPage } from './UploadPage';
+export { AnalysisPage } from './AnalysisPage';
+export { CutsPage } from './CutsPage';
+export { AudioPage } from './AudioPage';
+export { WorkspacePage } from './WorkspacePage';
+export { ExportPage } from './ExportPage';